build(version.py): move into python kotekan module
use it in rfi_receiver script
