+++ lib/version/make_version.cmake.in
@@ -44,7 +44,7 @@ execute_process(
 )
 
 execute_process(
-  COMMAND python3 ${SOURCE_DIR}/scripts/version.py
+  COMMAND python3 ${SOURCE_DIR}/python/kotekan/version.py
   WORKING_DIRECTORY ${SOURCE_DIR}
   OUTPUT_VARIABLE KOTEKAN_VERSION
   OUTPUT_STRIP_TRAILING_WHITESPACE
