+++ lib/version/version.h
@@ -16,7 +16,7 @@ extern "C" {
 
 /**
  * @brief Get the kotekan version string
- * See scripts/version.py to see how this is generated
+ * See python/kotekan/version.py to see how this is generated
  * This version is updated every time make is called
  *
  * @return The kotekan version string
