+++ python/setup.py
@@ -8,7 +8,7 @@ from future.builtins.disabled import *  # noqa  pylint: disable=W0401, W0614
 from os import path
 from setuptools import setup, find_packages
 
-import kotekan
+from kotekan.version import get_version
 
 here = path.abspath(path.dirname(__file__))
 
@@ -22,7 +22,7 @@ with open(path.join(here, "requirements.txt"), "r") as f:
 
 setup(
     name="kotekan",
-    version=kotekan.__version__,
+    version=get_version(),
     license="MIT",
     author="Kotekan Developers",
     description="Python support code for kotekan",
